created BookingViewSet

+++ littlelemon/restaurant/models.py
@@ -5,10 +5,10 @@ from django.db import models
 
 class Booking(models.Model):
     name = models.CharField(max_length=255)
-    no_of_guests = models.SmallIntegerField()
+    no_of_guests = models.IntegerField()
     boking_date = models.DateField(db_index=True)
 
 class Menu(models.Model):
     title = models.CharField(max_length=255)
     price = models.DecimalField(max_digits=10, decimal_places=2)
-    inventory = models.SmallIntegerField()
+    inventory = models.IntegerField()
